Add solution to diamond text pattern in Python
The str.center method makes things much cleaner. Obviously doable without
it, but it's perfect for this situation.

Otherwise, I like solving half of it and then adding the reverse. Maybe
a bit too much with the ternary cases for dealing with the A special
cases, nice and condensed though.