Rewrite Clj all-your-base with same algo as Python
The positional notation stuff made this way more complicated than it
needed to be. It was easy enough to convert the input digits into base10
with `map-indexed`, but trying to invert that operation was gnarly.
Because the power you raise each digit on depends on the number of
digits, I had `biggest-power-of-base` to figure out the number of
digits, then the `loop`+`recur` to go through each step.

But it turns out there's a much simpler way! Rather than mess with any
of exponents, just repeatedly multiply or divide by the base. Which is
equivalent but far simpler for a programmatic algorithm.