Python solution for Minesweeper
Apologies for the mega long lines, I need to figure out what Python
allows for splitting statements across lines to make that better.

I had an encounter with a matrix problem via LeetCode, so I had an idea
of what I wanted to do here. Since we need to process each square, the
nested loop is clear enough, but I don't want to mess with +1/-1 stuff
to check the surrounding squares. So instead, making a list of the
surrounding 8 squares, a bounds check filter, and count the number
of *s.

A couple tricky parts: the any/all for error checking, and converting
the row string into and out of a mutable list. I find some of the
behaviour of any/all tricky for handling their edge case tests, like I
know I want to say "if there are any characters that aren't '*' or ' ',
raise an exception" but it passes if `minefield` is an empty list, or if
the list contains an empty string. I guess because the list expression
generates nothing, then there's no explicit failure to hit the `all`.
Anyway, point being, that second exception check with three clauses
isn't ideal.

I have a few ideas for improvements based on the deep dive video, but
wanted to commit the first pass :D