Move more logic into nested loop in Minesweper
I like throwing the exception while iterating, we already have to check
every character so the `all` was extra work.

In the deep dive video on the problem, a few solutions had a structure
like:

[[-1, -1], [-1, 0], [1, 1]]

etc for the eight adjacent tiles. Which is what I was doing with the big
pile of zips and ranges, but clearer. I swapped that for a nested
list expression, then a check to ignore the (0, 0) case.

The `mine_neighbours or " "` bit is much nicer than the ternary I had
previously, as well.

Only thing I find a bit worse about this one is the giant if that
simulatenously checks boundaries, ignores the chosen tile, and then
finally checks for a mine. It would probably be better to split it up,
but I'm gonna leave it for now and work on other problems first.