Add solutions for two more resistor-color problems
I like the way the way the four resistance problems build on each other.

For the expert one, I had a nice way of doing it with negative indexing
to decide the range, rather than using unpacking like what I'm
committing now. But it broke the `"".join` part for the one transistor
case, so I had a special case to just return 0 ohms, and that seemed
lame. Checking how others handled it reminded me about the `*`
unpacking, which is perfect for this. I'm just so used to thinking in
slices, I guess.