Solutions for isogram and triangle in Clojure
Isogram is pretty trivial, but I like the cycle thing in triangle. When
I wrote out the needed comparisons as the index, I could see that there
was a pattern:

(1, 2) vs 0
(0, 2) vs 1
(0, 1) vs 2

So I did a little test of `(take 3 (cycle sides))`, `(take
3 (rest (cycle sides)))`, etc and it came out as the right pattern :)
Well, technically you get (2, 0) rather than (0, 2) but + is commutative
so it works out.