Add almost-correct Minesweeper solution in Clojure
This passes all the test cases except horizontal-line and vertical-line.
In the end I found it easier to ditch the min/max range checking
approach and instead put more conditinoals on the `for` loop. But I
thought it would be instructive to commit both versions.

To be honest, I'm actually surprised this one gets as close as it does
to working! It has some major issues with its bounds checking, and the
"line" tests expose that best. But I'm surprised none of the others had
issues.